pyproject.toml
[tool.poetry]
name = "lambda-workshop"
version = "0.1.0"
description = ""
authors = ["nikolasschriefer <info@nikolasschriefer.de>"]

[tool.poetry.dependencies]
python = "^3.9"
yahooquery = "^2.2"
pandas = "^1.4"
google-api-python-client = "^2.37"

[tool.poetry.dev-dependencies]
flake8 = "^4.0"
black = "^22.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
